Enforcing Barr C standard

+++ src/project_utils.h
@@ -13,7 +13,8 @@ struct reading_structure_t
     float r_m2;
 };
 
-enum event_e : uint8_t{
+enum event_e : uint8_t
+{
   EVENT_NONE,
   EVENT_HALL,
   EVENT_INPUT1,
@@ -26,10 +27,10 @@ enum event_e : uint8_t{
 boolean check_i2c();
 float mapfloat(float x, float in_min, float in_max, float out_min, float out_max);
 uint32_t get_bits(float x, float min, float max, int precision);
-void push_value(float value,reading_structure_t *values);
-float get_variance(reading_structure_t *values);
-void clear_value(reading_structure_t *values);
-unsigned int bcd2bin(unsigned char val);
-char bin2bcd(unsigned int val);
+void push_value(float value,reading_structure_t * values);
+float get_variance(reading_structure_t * values);
+void clear_value(reading_structure_t * values);
+uint16_t bcd2bin(unsigned char val);
+char bin2bcd(uint16_t val);
 
 #endif /* PROJECT_UTILS_H_ */
